Crate follow_chains

This example shows how to use the chain follower to follow all chains, until they have all reached tip. It will report on how many blocks for each chain exist between eras, and also how long each chain took to reach its tip.

Constants

RUNNING_UPDATE_INTERVAL 🔒

The interval between showing a block, even if nothing else changed.

Functions

follow_for 🔒

Try and follow a chain continuously, from Genesis until Tip.

log_bad_cip36_info 🔒

Helper function for logging bad CIP36.

log_cip509_info 🔒

Helper function for logging CIP509 validation.

main 🔒

process_argument 🔒

Process our CLI Arguments

start_sync_for 🔒

Start syncing a particular network

update_biggest_aux_data 🔒

Helper function for updating the biggest aux data.